+++ Data/SBV/TP.hs
@@ -48,7 +48,7 @@ module Data.SBV.TP (
        , sorry
 
        -- * Running TP proofs
-       , TP, runTP, runTPWith, tpQuiet, tpRibbon, tpStats, tpAsms
+       , TP, runTP, runTPWith, tpQuiet, tpStats, tpAsms
 
        -- * Measure helpers for smtFunctionWithMeasure
        , measureLemma, measureLemmaWith