+++ Documentation/SBV/Examples/TP/QuickSort.hs
@@ -329,7 +329,7 @@ partitionSndBound = sInduct "partitionNotLongerSnd"
 -- Functions proven terminating: count, lge, llt, nonDecreasing, partition, quickSort
 -- [Proven] quickSortIsCorrect :: Ɐxs ∷ [Integer] → Bool
 correctness :: forall a. (Eq a, OrdSymbolic (SBV a), SymVal a) => IO (Proof (Forall "xs" [a] -> SBool))
-correctness = runTPWith (tpRibbon 60 z3) $ do
+correctness = runTP $ do
 
   --------------------------------------------------------------------------------------------
   -- Part I. Import helper lemmas, definitions